% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{slides_gif}
\alias{slides_gif}
\title{Title}
\usage{
slides_gif(input, output = paste0(gsub("html$", "", input), "gif"),
  density = 200, delay = 0.5)
}
\arguments{
\item{input}{name of the html slides file}

\item{output}{name of the gif}

\item{density}{resolution maybe}

\item{delay}{time frame is displayed in seconds}
}
\value{
a pdf and a gif
}
\description{
Title
}
\examples{

}
